CONSOLE-4125,CONSOLE-4294: Rename user preferences PerspectiveDropdown to PreferredPerspectiveSelect
