+++ frontend/packages/console-app/src/components/user-preferences/perspective/__tests__/PreferredPerspectiveSelect.spec.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Select as SelectDeprecated } from '@patternfly/react-core/deprecated';
 import { shallow, ShallowWrapper } from 'enzyme';
 import { useExtensions } from '@console/plugin-sdk/src';
-import PerspectiveDropdown from '../PerspectiveDropdown';
+import PreferredPerspectiveSelect from '../PreferredPerspectiveSelect';
 import { usePreferredPerspective } from '../usePreferredPerspective';
 import { mockPerspectiveExtensions } from './perspective.data';
 
@@ -35,7 +35,7 @@ describe('PerspectiveDropdown', () => {
   it('should render skeleton if user preferences have not loaded', () => {
     useExtensionsMock.mockReturnValue(mockPerspectiveExtensions);
     usePreferredPerspectiveMock.mockReturnValue(['', jest.fn(), false]);
-    wrapper = shallow(<PerspectiveDropdown />);
+    wrapper = shallow(<PreferredPerspectiveSelect />);
     expect(
       wrapper.find('[data-test="dropdown skeleton console.preferredPerspective"]').exists(),
     ).toBeTruthy();
@@ -44,7 +44,7 @@ describe('PerspectiveDropdown', () => {
   it('should render select with value corresponding to preferred perspective if user preferences have loaded and preferred perspective is defined', () => {
     useExtensionsMock.mockReturnValue(mockPerspectiveExtensions);
     usePreferredPerspectiveMock.mockReturnValue([preferredPerspectiveValue, jest.fn(), true]);
-    wrapper = shallow(<PerspectiveDropdown />);
+    wrapper = shallow(<PreferredPerspectiveSelect />);
     expect(
       wrapper.find('[data-test="dropdown console.preferredPerspective"]').exists(),
     ).toBeTruthy();
@@ -54,7 +54,7 @@ describe('PerspectiveDropdown', () => {
   it('should render select with value "Last viewed" if user preferences have loaded but preferred perspective is not defined', () => {
     useExtensionsMock.mockReturnValue(mockPerspectiveExtensions);
     usePreferredPerspectiveMock.mockReturnValue([undefined, jest.fn(), true]);
-    wrapper = shallow(<PerspectiveDropdown />);
+    wrapper = shallow(<PreferredPerspectiveSelect />);
     expect(
       wrapper.find('[data-test="dropdown console.preferredPerspective"]').exists(),
     ).toBeTruthy();
